update and refactor code
