+++ manager/src/main/java/org/dromara/hertzbeat/manager/scheduler/netty/process/HeartbeatProcessor.java
@@ -1,9 +1,9 @@
-package org.dromara.hertzbeat.manager.netty.process;
+package org.dromara.hertzbeat.manager.scheduler.netty.process;
 
 import io.netty.channel.ChannelHandlerContext;
 import lombok.extern.slf4j.Slf4j;
 import org.dromara.hertzbeat.common.entity.message.ClusterMsg;
-import org.dromara.hertzbeat.manager.netty.ManageServer;
+import org.dromara.hertzbeat.manager.scheduler.netty.ManageServer;
 import org.dromara.hertzbeat.remoting.netty.NettyRemotingProcessor;
 
 /**
